graph.c: Always apply pending changes rather than checking.
Checking for pending changes is highly non-portable, particularly to
offloaded / accelerated GraphBLAS implementations.  It's simpler to rely
on the implementation being sensibly quick about applying pending
changes in the critical section.

Alternately, we could assume that nvals(m) (to be replaced by
GrB_wait(m) in a future spec version) is thread-safe and move it outside
the critical section.