In Graph_GetNodeEdges, use GrB_Col_extract rather than an iterator.
This ends up being approximately the same as a portable iterator
implementation and should assist any GraphBLAS that keeps the graph in
another memory space (local to a GPU, FPGA, etc.).

The edge_accum user-defined function also causes issues for offloading,
so that is replaced with a typical loop on the non-offloaded side.  That
also is where the edge array entries point; those pointers are not to
data on the offloaded side.